[feat/#175] :recycle: Refactor: PlayersStatus를 일급컬렉션 도메인과 DTO를 따로 분리
