+++ chatty-be/src/main/java/com/chatty/chatty/quizroom/repository/PlayersStatusRepository.java
@@ -1,8 +1,9 @@
 package com.chatty.chatty.quizroom.repository;
 
-import com.chatty.chatty.quizroom.controller.dto.PlayerStatus;
+import com.chatty.chatty.quizroom.domain.PlayerStatus;
 import com.chatty.chatty.quizroom.domain.PlayersStatus;
 import java.util.Map;
+import java.util.Optional;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 import lombok.RequiredArgsConstructor;
@@ -16,21 +17,17 @@ public class RoomUsersStatusRepository {
 
     private static final Map<Long, PlayersStatus> roomUsersStatuses = new ConcurrentHashMap<>();
 
-    public PlayersStatus findByRoomId(Long roomId) {
+    public Optional<PlayersStatus> findByRoomId(Long roomId) {
         return roomUsersStatuses.get(roomId);
     }
 
     public PlayersStatus addUserToRoom(Long roomId, Long userId) {
-        return roomUsersStatuses.computeIfPresent(
-                roomId, (id, existingStatus) -> existingStatus.join(userId)
-        );
+
     }
 
-    public void initializeRoomWithManager(Long roomId, Long roomManagerId) {
-        PlayersStatus newStatus = PlayersStatus.builder()
-                .roomId(roomId)
-                .roomUserStatuses(Set.of(PlayerStatus.createStatus(roomManagerId)))
+    public PlayersStatus initRoom(Long roomId) {
+        return PlayersStatus.builder()
+                .playerStatuses(Set.of())
                 .build();
-        roomUsersStatuses.put(roomId, newStatus);
     }
 }
